Secured views

+++ products/views.py
@@ -2,6 +2,7 @@ from functools import reduce
 import operator
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category, Manufacturer, TyreSize
@@ -9,6 +10,7 @@ from .models import Product, Category, Manufacturer, TyreSize
 from .forms import ProductForm
 
 # Create your views here.
+
 
 def all_products(request):
     """ A view to return all products page """
@@ -26,8 +28,6 @@ def all_products(request):
 
     sort = None
     direction = None
-
-
 
     if request.GET:
         # Handles free search bar functionality
@@ -169,8 +169,12 @@ def product_details(request, ean_code):
     return render(request, 'products/product_details.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, You don't have permission to acces this page!")
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -192,8 +196,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def update_product(request, ean_code):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, You don't have permission to acces this page!")
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, ean_code=ean_code)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -217,10 +226,14 @@ def update_product(request, ean_code):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, ean_code):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, You don't have permission to acces this page!")
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, ean_code=ean_code)
     product.delete()
     messages.success(request, 'Product deleted succcessfully!')
     return redirect(reverse('products'))
-
